+++ src/utils.h
@@ -24,9 +24,8 @@
 
 #include <QtGlobal>
 
-namespace Utils {
+namespace Utils
+{
     static inline float cleanerRatio(qint64 sizeBefore, qint64 sizeAfter)
-    {
-        return 100.0f - float(sizeAfter) / sizeBefore * 100.0f;
-    }
+    { return 100.0f - float(sizeAfter) / sizeBefore * 100.0f; }
 }